+++ src/BDSBunchHalo.cc
@@ -62,9 +62,9 @@ void  BDSBunchHalo::SetOptions(const BDSParticleDefinition* beamParticle,
 			       const GMAD::Beam& beam,
 			       const BDSBunchType& distrType,
 			       G4Transform3D beamlineTransformIn,
-			       const G4double beamlineS)
+			       const G4double beamlineSIn)
 {
-  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineS);
+  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineSIn);
   alphaX                = G4double(beam.alfx);
   alphaY                = G4double(beam.alfy);
   betaX                 = G4double(beam.betx);