+++ src/BDSBunchRing.cc
@@ -45,9 +45,9 @@ void BDSBunchRing::SetOptions(const BDSParticleDefinition* beamParticle,
 			      const GMAD::Beam& beam,
 			      const BDSBunchType& distrType,
 			      G4Transform3D beamlineTransformIn,
-			      const G4double beamlineS)
+			      const G4double beamlineSIn)
 {
-  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineS);
+  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineSIn);
   rMin = beam.Rmin * CLHEP::m;
   rMax = beam.Rmax * CLHEP::m;
   rDif = rMax - rMin;