+++ src/BDSBunchSquare.cc
@@ -45,9 +45,9 @@ void BDSBunchSquare::SetOptions(const BDSParticleDefinition* beamParticle,
 				const GMAD::Beam& beam,
 				const BDSBunchType& distrType,
 				G4Transform3D beamlineTransformIn,
-				const G4double beamlineS)
+				const G4double beamlineSIn)
 {
-  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineS);
+  BDSBunch::SetOptions(beamParticle, beam, distrType, beamlineTransformIn, beamlineSIn);
   envelopeX  = beam.envelopeX  * CLHEP::m;
   envelopeY  = beam.envelopeY  * CLHEP::m;
   envelopeXp = beam.envelopeXp * CLHEP::rad;