+++ manual/source/version_history.rst
@@ -14,9 +14,16 @@ Bug Fixes
 ---------
 
 * Fixed reloading user file when reading more particles than defined in the file.
+* Fixed infinite tracking from nans return from field map when BDSIM format
+  field map file was lacking lower and upper limits.
 
 Output Changes
 --------------
+
+* Memory usage (for Mac & Linux) added at the end of each event in event info. This
+  is the memory usage of the whole program at that point including event independent
+  quantities such as the model.
+* BDSOutputROOTEventInfo class version incremented to 4.
 
 Utilities
 ---------