+++ manual/source/version_history.rst
@@ -1,9 +1,10 @@
-V1.3 - 2018 / ?? / ??
+V1.3 - 2018 / 09 / ??
 =====================
 
 New Features
 ------------
 
+* All Geant4 reference physics lists are now available.
 * New beam pipe aperture for the CLIC post collision line.
 
 General
@@ -12,8 +13,17 @@ General
 Bug Fixes
 ---------
 
+* Fixed reloading user file when reading more particles than defined in the file.
+* Fixed infinite tracking from nans return from field map when BDSIM format
+  field map file was lacking lower and upper limits.
+
 Output Changes
 --------------
+
+* Memory usage (for Mac & Linux) added at the end of each event in event info. This
+  is the memory usage of the whole program at that point including event independent
+  quantities such as the model.
+* BDSOutputROOTEventInfo class version incremented to 4.
 
 Utilities
 ---------